Some enhancements because of new version of epm
