+++ src/main/java/com/jasonwjones/pbcs/api/v3/JobLaunchPayload.java
@@ -4,47 +4,24 @@ import java.util.Map;
 
 import com.fasterxml.jackson.annotation.JsonValue;
 
-public class JobLaunchPayload {
+/***
+ * Some of epm requests need to have json like:
+ *{
+ * 	"backupName": "2022-02-16T21:00:02/Artifact_Snapshot_2021-12-16T21:00:02",
+ * 	"parameters": {
+ * 		"targetName": "Backup_16Dec"
+ *  }
+ * }
+ * Need have map of parameters directly in the body. For requests which do not need this use @{@link MetadataImportPayload}
+ */
+public class JobLaunchPayload extends Payload {
 
-	private String jobType;
-	
-	private String jobName;
-
-	private Map<String, String> parameters;
-
-	public JobLaunchPayload() {
-	}
-	
 	public JobLaunchPayload(String jobType, String jobName) {
-		this.jobType = jobType;
-		this.jobName = jobName;
-	}
-
-	public String getJobType() {
-		return jobType;
-	}
-
-	public void setJobType(String jobType) {
-		this.jobType = jobType;
-	}
-
-	public String getJobName() {
-		return jobName;
-	}
-
-	public void setJobName(String jobName) {
-		this.jobName = jobName;
+		super(jobType, jobName);
 	}
 
 	@JsonValue
 	public Map<String, String> getParameters() {
-		return parameters;
+		return super.getParameters();
 	}
-
-	public void setParameters(Map<String, String> parameters) {
-		this.parameters = parameters;
-		this.parameters.put("jobType", jobType);
-		this.parameters.put("jobName", jobName);
-	}
-	
 }
